Initial commit
This is a port of the original pysk8 driver at
https://github.com/andrewramsay/sk8-drivers/tree/master/pysk8.

The original version of pysk8 relied on a specific BLE dongle which
communicated with the OS over a serial interface. This works but obviously
places extra constraints on the use of SK8s as they should be compatible
with any host device that supports BLE.

To provide a more generic implementation, this version of the code is
based around the cross-platform Bleak library which provides a common API
to multiple Bluetooth backends (https://github.com/hbldh/bleak). The
pysk8 API remains very similar to the original version but should now be
usable with any BLE-capable hardware.

Current status: basic functionality is working, including streaming of
sensor data. Still need to test all the API methods work, add some
automated tests, and port over the example scripts.